added template to serializer
